@@ -1,3 +1,5 @@
 module github.com/afadesigns/zshellcheck
 
-go 1.22.5
+go 1.23.0
+
+require gopkg.in/yaml.v3 v3.0.1
