@@ -1,9 +1,9 @@
 package main
 
 import (
+	"flag"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 
 	"github.com/afadesigns/zshellcheck/pkg/ast"
@@ -11,21 +11,56 @@ import (
 	"github.com/afadesigns/zshellcheck/pkg/lexer"
 	"github.com/afadesigns/zshellcheck/pkg/parser"
 	"github.com/afadesigns/zshellcheck/pkg/reporter"
+	"gopkg.in/yaml.v3"
 )
 
+type Config struct {
+	DisabledKatas []string `yaml:"disabled_katas"`
+}
+
 func main() {
-	if len(os.Args) < 2 {
-		fmt.Println("Usage: zshellcheck <file1.zsh> [file2.zsh]...")
+	format := flag.String("format", "text", "The output format (text or json)")
+	flag.Parse()
+
+	if len(flag.Args()) < 1 {
+		fmt.Println("Usage: zshellcheck [flags] <file1.zsh> [file2.zsh]...")
 		os.Exit(1)
 	}
 
-	for _, filename := range os.Args[1:] {
-		processFile(filename, os.Stdout, os.Stderr)
+	config, err := loadConfig(".zshellcheckrc")
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error loading config: %s\n", err)
+		os.Exit(1)
+	}
+
+	kataRegistry := &katas.Registry
+
+	for _, filename := range flag.Args() {
+		processFile(filename, os.Stdout, os.Stderr, config, kataRegistry, *format)
 	}
 }
 
-func processFile(filename string, out, errOut io.Writer) {
-	data, err := ioutil.ReadFile(filename)
+func loadConfig(path string) (Config, error) {
+	var config Config
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		return config, nil
+	}
+
+	data, err := os.ReadFile(path)
+	if err != nil {
+		return config, err
+	}
+
+	err = yaml.Unmarshal(data, &config)
+	if err != nil {
+		return config, err
+	}
+
+	return config, nil
+}
+
+func processFile(filename string, out, errOut io.Writer, config Config, registry *katas.KatasRegistry, format string) {
+	data, err := os.ReadFile(filename)
 	if err != nil {
 		fmt.Fprintf(errOut, "Error reading file %s: %s\n", filename, err)
 		return
@@ -45,13 +80,21 @@ func processFile(filename string, out, errOut io.Writer) {
 
 	violations := []katas.Violation{}
 	ast.Walk(program, func(node ast.Node) bool {
-		violations = append(violations, katas.Check(node)...)
+		violations = append(violations, registry.Check(node, config.DisabledKatas)...)
 		return true // Continue walking
 	})
 
 	if len(violations) > 0 {
-		fmt.Fprintf(out, "Violations in %s:\n", filename)
-		reporter := reporter.NewTextReporter(out)
-		reporter.Report(violations)
+		var r reporter.Reporter
+		switch format {
+		case "json":
+			r = reporter.NewJSONReporter(out)
+		default:
+			fmt.Fprintf(out, "Violations in %s:\n", filename)
+			r = reporter.NewTextReporter(out)
+		}
+		if err := r.Report(violations); err != nil {
+			fmt.Fprintf(errOut, "Error reporting violations: %s\n", err)
+		}
 	}
 }
