@@ -6,9 +6,106 @@ import (
 	"github.com/afadesigns/zshellcheck/pkg/token"
 )
 
+type NodeType int
+
+const (
+	ProgramNode NodeType = iota
+	LetStatementNode
+	ReturnStatementNode
+	ExpressionStatementNode
+	IdentifierNode
+	IntegerLiteralNode
+	BooleanNode
+	PrefixExpressionNode
+	PostfixExpressionNode
+	InfixExpressionNode
+	BlockStatementNode
+	IfStatementNode
+	ForLoopStatementNode
+	WhileLoopStatementNode
+	FunctionLiteralNode
+	CallExpressionNode
+	StringLiteralNode
+	BracketExpressionNode
+	DoubleBracketExpressionNode
+	ArrayAccessNode
+	InvalidArrayAccessNode
+	CommandSubstitutionNode
+	ShebangNode
+	DollarParenExpressionNode
+	SimpleCommandNode
+	IndexExpressionNode
+	ConcatenatedExpressionNode
+	CaseStatementNode
+	RedirectionNode
+	FunctionDefinitionNode
+	GroupedExpressionNode
+)
+
+type GroupedExpression struct {
+	Token token.Token // The '(' token
+	Exp   Expression
+}
+
+func (ge *GroupedExpression) Type() NodeType                { return GroupedExpressionNode }
+func (ge *GroupedExpression) expressionNode()               {}
+func (ge *GroupedExpression) TokenLiteral() string          { return ge.Token.Literal }
+func (ge *GroupedExpression) TokenLiteralNode() token.Token { return ge.Token }
+func (ge *GroupedExpression) String() string {
+	return "(" + ge.Exp.String() + ")"
+}
+
+type FunctionDefinition struct {
+	Token token.Token // The name token
+	Name  *Identifier
+	Body  Statement // The function body (usually BlockStatement)
+}
+
+func (fd *FunctionDefinition) Type() NodeType       { return FunctionDefinitionNode }
+func (fd *FunctionDefinition) expressionNode()      {}
+func (fd *FunctionDefinition) statementNode()       {}
+func (fd *FunctionDefinition) TokenLiteral() string { return fd.Token.Literal }
+func (fd *FunctionDefinition) String() string {
+	var out []byte
+	if fd.Name != nil {
+		out = append(out, []byte(fd.Name.String())...)
+	}
+	out = append(out, []byte("() ")...)
+	if fd.Body != nil {
+		out = append(out, []byte(fd.Body.String())...)
+	}
+	return string(out)
+}
+
+type Redirection struct {
+	Token    token.Token
+	Left     Expression
+	Operator string
+	Right    Expression
+}
+
+func (r *Redirection) Type() NodeType       { return RedirectionNode }
+func (r *Redirection) expressionNode()      {}
+func (r *Redirection) TokenLiteral() string { return r.Token.Literal }
+func (r *Redirection) String() string {
+	var out []byte
+	if r.Left != nil {
+		out = append(out, []byte(r.Left.String())...)
+	}
+	out = append(out, []byte(" ")...)
+	out = append(out, []byte(r.Operator)...)
+	out = append(out, []byte(" ")...)
+	if r.Right != nil {
+		out = append(out, []byte(r.Right.String())...)
+	}
+	return string(out)
+}
+
 type Node interface {
 	TokenLiteral() string
 	String() string
+	Type() NodeType
+	TokenLiteralNode() token.Token
 }
 
 type Statement interface {
@@ -25,11 +122,19 @@ type Program struct {
 	Statements []Statement
 }
 
+func (p *Program) Type() NodeType { return ProgramNode }
 func (p *Program) TokenLiteral() string {
 	if len(p.Statements) > 0 {
 		return p.Statements[0].TokenLiteral()
 	}
 	return ""
+}
+
+func (p *Program) TokenLiteralNode() token.Token {
+	if len(p.Statements) > 0 {
+		return p.Statements[0].TokenLiteralNode()
+	}
+	return token.Token{}
 }
 
 func (p *Program) String() string {
@@ -46,6 +151,7 @@ type LetStatement struct {
 	Value Expression
 }
 
+func (ls *LetStatement) Type() NodeType       { return LetStatementNode }
 func (ls *LetStatement) statementNode()       {}
 func (ls *LetStatement) TokenLiteral() string { return ls.Token.Literal }
 func (ls *LetStatement) String() string {
@@ -66,6 +172,7 @@ type ReturnStatement struct {
 	ReturnValue Expression
 }
 
+func (rs *ReturnStatement) Type() NodeType       { return ReturnStatementNode }
 func (rs *ReturnStatement) statementNode()       {}
 func (rs *ReturnStatement) TokenLiteral() string { return rs.Token.Literal }
 func (rs *ReturnStatement) String() string {
@@ -84,6 +191,7 @@ type ExpressionStatement struct {
 	Expression Expression
 }
 
+func (es *ExpressionStatement) Type() NodeType       { return ExpressionStatementNode }
 func (es *ExpressionStatement) statementNode()       {}
 func (es *ExpressionStatement) TokenLiteral() string { return es.Token.Literal }
 func (es *ExpressionStatement) String() string {
@@ -98,6 +206,7 @@ type Identifier struct {
 	Value string
 }
 
+func (i *Identifier) Type() NodeType       { return IdentifierNode }
 func (i *Identifier) expressionNode()      {}
 func (i *Identifier) TokenLiteral() string { return i.Token.Literal }
 func (i *Identifier) String() string       { return i.Value }
@@ -107,6 +216,7 @@ type IntegerLiteral struct {
 	Value int64
 }
 
+func (il *IntegerLiteral) Type() NodeType       { return IntegerLiteralNode }
 func (il *IntegerLiteral) expressionNode()      {}
 func (il *IntegerLiteral) TokenLiteral() string { return il.Token.Literal }
 func (il *IntegerLiteral) String() string       { return il.Token.Literal }
@@ -116,6 +226,7 @@ type Boolean struct {
 	Value bool
 }
 
+func (b *Boolean) Type() NodeType       { return BooleanNode }
 func (b *Boolean) expressionNode()      {}
 func (b *Boolean) TokenLiteral() string { return b.Token.Literal }
 func (b *Boolean) String() string       { return b.Token.Literal }
@@ -126,6 +237,7 @@ type PrefixExpression struct {
 	Right    Expression
 }
 
+func (pe *PrefixExpression) Type() NodeType       { return PrefixExpressionNode }
 func (pe *PrefixExpression) expressionNode()      {}
 func (pe *PrefixExpression) TokenLiteral() string { return pe.Token.Literal }
 func (pe *PrefixExpression) String() string {
@@ -145,6 +257,7 @@ type PostfixExpression struct {
 	Operator string
 }
 
+func (pe *PostfixExpression) Type() NodeType       { return PostfixExpressionNode }
 func (pe *PostfixExpression) expressionNode()      {}
 func (pe *PostfixExpression) TokenLiteral() string { return pe.Token.Literal }
 func (pe *PostfixExpression) String() string {
@@ -165,6 +278,7 @@ type InfixExpression struct {
 	Right    Expression
 }
 
+func (ie *InfixExpression) Type() NodeType       { return InfixExpressionNode }
 func (ie *InfixExpression) expressionNode()      {}
 func (ie *InfixExpression) TokenLiteral() string { return ie.Token.Literal }
 func (ie *InfixExpression) String() string {
@@ -188,6 +302,7 @@ type BlockStatement struct {
 	Statements []Statement
 }
 
+func (bs *BlockStatement) Type() NodeType       { return BlockStatementNode }
 func (bs *BlockStatement) statementNode()       {}
 func (bs *BlockStatement) TokenLiteral() string { return bs.Token.Literal }
 func (bs *BlockStatement) String() string {
@@ -200,11 +315,12 @@ func (bs *BlockStatement) String() string {
 
 type IfStatement struct {
 	Token       token.Token // The 'if' token
-	Condition   Expression
+	Condition   *BlockStatement
 	Consequence *BlockStatement
 	Alternative *BlockStatement
 }
 
+func (is *IfStatement) Type() NodeType       { return IfStatementNode }
 func (is *IfStatement) statementNode()       {}
 func (is *IfStatement) TokenLiteral() string { return is.Token.Literal }
 func (is *IfStatement) String() string {
@@ -226,30 +342,53 @@ func (is *IfStatement) String() string {
 }
 
 type ForLoopStatement struct {
-	Token     token.Token // The 'for' token
-	Init      Expression
-	Condition Expression
-	Post      Expression
+	Token     token.Token  // The 'for' token
+	Init      Expression   // Variable name (for-each) or Init expr (arithmetic)
+	Condition Expression   // Arithmetic condition
+	Post      Expression   // Arithmetic post
+	Items     []Expression // Items to iterate over (for-each)
 	Body      *BlockStatement
 }
 
+func (fls *ForLoopStatement) Type() NodeType       { return ForLoopStatementNode }
 func (fls *ForLoopStatement) statementNode()       {}
 func (fls *ForLoopStatement) TokenLiteral() string { return fls.Token.Literal }
 func (fls *ForLoopStatement) String() string {
 	var out []byte
-	out = append(out, []byte("for ((")...)
-	if fls.Init != nil {
-		out = append(out, []byte(fls.Init.String())...)
+	// Heuristic: if Condition or Post is present, or Items is nil (and not implicit?), it's arithmetic?
+	// Actually, explicit `Items` makes it for-each.
+	// But `for i` (implicit in) has Items=nil.
+	// Arithmetic `for ((...))` usually has params. `for ((;;))` is possible.
+	// I'll assume if Items is non-nil (even empty) it's for-each.
+	// Or if Init is Identifier and others are nil?
+
+	if fls.Items != nil {
+		out = append(out, []byte("for ")...)
+		if fls.Init != nil {
+			out = append(out, []byte(fls.Init.String())...)
+		}
+		out = append(out, []byte(" in ")...)
+		for _, item := range fls.Items {
+			out = append(out, []byte(item.String())...)
+			out = append(out, []byte(" ")...)
+		}
+		out = append(out, []byte("; do ")...)
+	} else {
+		out = append(out, []byte("for ((")...)
+		if fls.Init != nil {
+			out = append(out, []byte(fls.Init.String())...)
+		}
+		out = append(out, []byte("; ")...)
+		if fls.Condition != nil {
+			out = append(out, []byte(fls.Condition.String())...)
+		}
+		out = append(out, []byte("; ")...)
+		if fls.Post != nil {
+			out = append(out, []byte(fls.Post.String())...)
+		}
+		out = append(out, []byte(")); do ")...)
 	}
-	out = append(out, []byte("; ")...)
-	if fls.Condition != nil {
-		out = append(out, []byte(fls.Condition.String())...)
-	}
-	out = append(out, []byte("; ")...)
-	if fls.Post != nil {
-		out = append(out, []byte(fls.Post.String())...)
-	}
-	out = append(out, []byte(")); do ")...)
+
 	if fls.Body != nil {
 		out = append(out, []byte(fls.Body.String())...)
 	}
@@ -257,12 +396,37 @@ func (fls *ForLoopStatement) String() string {
 	return string(out)
 }
 
+type WhileLoopStatement struct {
+	Token     token.Token // The 'while' token
+	Condition *BlockStatement
+	Body      *BlockStatement
+}
+
+func (wls *WhileLoopStatement) Type() NodeType       { return WhileLoopStatementNode }
+func (wls *WhileLoopStatement) statementNode()       {}
+func (wls *WhileLoopStatement) TokenLiteral() string { return wls.Token.Literal }
+func (wls *WhileLoopStatement) String() string {
+	var out []byte
+	out = append(out, []byte("while ")...)
+	if wls.Condition != nil {
+		out = append(out, []byte(wls.Condition.String())...)
+	}
+	out = append(out, []byte("; do ")...)
+	if wls.Body != nil {
+		out = append(out, []byte(wls.Body.String())...)
+	}
+	out = append(out, []byte("done")...)
+	return string(out)
+}
+
 type FunctionLiteral struct {
 	Token      token.Token // The 'fn' token
+	Name       string      // The function name (optional)
 	Parameters []*Identifier
 	Body       *BlockStatement
 }
 
+func (fl *FunctionLiteral) Type() NodeType       { return FunctionLiteralNode }
 func (fl *FunctionLiteral) expressionNode()      {}
 func (fl *FunctionLiteral) TokenLiteral() string { return fl.Token.Literal }
 func (fl *FunctionLiteral) String() string {
@@ -286,6 +450,7 @@ type CallExpression struct {
 	Arguments []Expression
 }
 
+func (ce *CallExpression) Type() NodeType       { return CallExpressionNode }
 func (ce *CallExpression) expressionNode()      {}
 func (ce *CallExpression) TokenLiteral() string { return ce.Token.Literal }
 func (ce *CallExpression) String() string {
@@ -306,6 +471,7 @@ type StringLiteral struct {
 	Value string
 }
 
+func (sl *StringLiteral) Type() NodeType       { return StringLiteralNode }
 func (sl *StringLiteral) expressionNode()      {}
 func (sl *StringLiteral) TokenLiteral() string { return sl.Token.Literal }
 func (sl *StringLiteral) String() string       { return sl.Token.Literal }
@@ -315,6 +481,7 @@ type BracketExpression struct {
 	Expressions []Expression
 }
 
+func (be *BracketExpression) Type() NodeType       { return BracketExpressionNode }
 func (be *BracketExpression) expressionNode()      {}
 func (be *BracketExpression) TokenLiteral() string { return be.Token.Literal }
 func (be *BracketExpression) String() string {
@@ -334,6 +501,7 @@ type DoubleBracketExpression struct {
 	Expressions []Expression
 }
 
+func (dbe *DoubleBracketExpression) Type() NodeType       { return DoubleBracketExpressionNode }
 func (dbe *DoubleBracketExpression) expressionNode()      {}
 func (dbe *DoubleBracketExpression) TokenLiteral() string { return dbe.Token.Literal }
 func (dbe *DoubleBracketExpression) String() string {
@@ -348,13 +516,13 @@ func (dbe *DoubleBracketExpression) String() string {
 	return string(out)
 }
 
-
 type ArrayAccess struct {
 	Token token.Token // The '${' token
 	Left  Expression
 	Index Expression
 }
 
+func (aa *ArrayAccess) Type() NodeType       { return ArrayAccessNode }
 func (aa *ArrayAccess) expressionNode()      {}
 func (aa *ArrayAccess) TokenLiteral() string { return aa.Token.Literal }
 func (aa *ArrayAccess) String() string {
@@ -367,12 +535,36 @@ func (aa *ArrayAccess) String() string {
 	return string(out)
 }
 
+type IndexExpression struct {
+	Token token.Token // The [ token
+	Left  Expression
+	Index Expression
+}
+
+func (ie *IndexExpression) Type() NodeType       { return IndexExpressionNode }
+func (ie *IndexExpression) expressionNode()      {}
+func (ie *IndexExpression) TokenLiteral() string { return ie.Token.Literal }
+func (ie *IndexExpression) String() string {
+	var out []byte
+	out = append(out, []byte("(")...)
+	if ie.Left != nil {
+		out = append(out, []byte(ie.Left.String())...)
+	}
+	out = append(out, []byte("[")...)
+	if ie.Index != nil {
+		out = append(out, []byte(ie.Index.String())...)
+	}
+	out = append(out, []byte("])")...)
+	return string(out)
+}
+
 type InvalidArrayAccess struct {
 	Token token.Token // The '$' token
 	Left  Expression
 	Index Expression
 }
 
+func (iaa *InvalidArrayAccess) Type() NodeType       { return InvalidArrayAccessNode }
 func (iaa *InvalidArrayAccess) expressionNode()      {}
 func (iaa *InvalidArrayAccess) TokenLiteral() string { return iaa.Token.Literal }
 func (iaa *InvalidArrayAccess) String() string {
@@ -390,6 +582,7 @@ type CommandSubstitution struct {
 	Command Expression
 }
 
+func (cs *CommandSubstitution) Type() NodeType       { return CommandSubstitutionNode }
 func (cs *CommandSubstitution) expressionNode()      {}
 func (cs *CommandSubstitution) TokenLiteral() string { return cs.Token.Literal }
 func (cs *CommandSubstitution) String() string {
@@ -401,6 +594,7 @@ type Shebang struct {
 	Path  string
 }
 
+func (s *Shebang) Type() NodeType       { return ShebangNode }
 func (s *Shebang) statementNode()       {}
 func (s *Shebang) TokenLiteral() string { return s.Token.Literal }
 func (s *Shebang) String() string {
@@ -412,6 +606,7 @@ type DollarParenExpression struct {
 	Command Expression
 }
 
+func (dpe *DollarParenExpression) Type() NodeType       { return DollarParenExpressionNode }
 func (dpe *DollarParenExpression) expressionNode()      {}
 func (dpe *DollarParenExpression) TokenLiteral() string { return dpe.Token.Literal }
 func (dpe *DollarParenExpression) String() string {
@@ -428,6 +623,7 @@ type SimpleCommand struct {
 	Arguments []Expression
 }
 
+func (sc *SimpleCommand) Type() NodeType       { return SimpleCommandNode }
 func (sc *SimpleCommand) expressionNode()      {}
 func (sc *SimpleCommand) TokenLiteral() string { return sc.Token.Literal }
 func (sc *SimpleCommand) String() string {
@@ -439,6 +635,69 @@ func (sc *SimpleCommand) String() string {
 	out = append(out, []byte(sc.Name.String())...)
 	out = append(out, []byte(" ")...)
 	out = append(out, []byte(strings.Join(args, " "))...)
+	return string(out)
+}
+
+type ConcatenatedExpression struct {
+	Token token.Token
+	Parts []Expression
+}
+
+func (ce *ConcatenatedExpression) Type() NodeType       { return ConcatenatedExpressionNode }
+func (ce *ConcatenatedExpression) expressionNode()      {}
+func (ce *ConcatenatedExpression) TokenLiteral() string { return ce.Token.Literal }
+func (ce *ConcatenatedExpression) String() string {
+	var out []byte
+	for _, p := range ce.Parts {
+		if p != nil {
+			out = append(out, []byte(p.String())...)
+		}
+	}
+	return string(out)
+}
+
+type CaseStatement struct {
+	Token   token.Token // The 'case' token
+	Value   Expression
+	Clauses []*CaseClause
+}
+
+func (cs *CaseStatement) Type() NodeType       { return CaseStatementNode }
+func (cs *CaseStatement) statementNode()       {}
+func (cs *CaseStatement) TokenLiteral() string { return cs.Token.Literal }
+func (cs *CaseStatement) String() string {
+	var out []byte
+	out = append(out, []byte("case ")...)
+	if cs.Value != nil {
+		out = append(out, []byte(cs.Value.String())...)
+	}
+	out = append(out, []byte(" in ")...)
+	for _, c := range cs.Clauses {
+		out = append(out, []byte(c.String())...)
+	}
+	out = append(out, []byte("esac")...)
+	return string(out)
+}
+
+type CaseClause struct {
+	Token    token.Token // The first token of pattern
+	Patterns []Expression
+	Body     *BlockStatement
+}
+
+func (cc *CaseClause) String() string {
+	var out []byte
+	for i, p := range cc.Patterns {
+		out = append(out, []byte(p.String())...)
+		if i < len(cc.Patterns)-1 {
+			out = append(out, []byte(" | ")...)
+		}
+	}
+	out = append(out, []byte(") ")...)
+	if cc.Body != nil {
+		out = append(out, []byte(cc.Body.String())...)
+	}
+	out = append(out, []byte(" ;; ")...)
 	return string(out)
 }
 
@@ -500,6 +759,12 @@ func Walk(node Node, f WalkFn) {
 		Walk(n.Init, f)
 		Walk(n.Condition, f)
 		Walk(n.Post, f)
+		for _, item := range n.Items {
+			Walk(item, f)
+		}
+		Walk(n.Body, f)
+	case *WhileLoopStatement:
+		Walk(n.Condition, f)
 		Walk(n.Body, f)
 	case *FunctionLiteral:
 		for _, param := range n.Parameters {
@@ -527,10 +792,14 @@ func Walk(node Node, f WalkFn) {
 	case *ArrayAccess:
 		Walk(n.Left, f)
 		Walk(n.Index, f)
+	case *IndexExpression:
+		Walk(n.Left, f)
+		Walk(n.Index, f)
 	case *InvalidArrayAccess:
 		Walk(n.Left, f)
 		Walk(n.Index, f)
 	case *CommandSubstitution:
+		Walk(n.Command, f)
 	case *DollarParenExpression:
 		Walk(n.Command, f)
 	case *SimpleCommand:
@@ -538,5 +807,56 @@ func Walk(node Node, f WalkFn) {
 		for _, arg := range n.Arguments {
 			Walk(arg, f)
 		}
+	case *ConcatenatedExpression:
+		for _, p := range n.Parts {
+			Walk(p, f)
+		}
+	case *CaseStatement:
+		Walk(n.Value, f)
+		for _, c := range n.Clauses {
+			for _, p := range c.Patterns {
+				Walk(p, f)
+			}
+			Walk(c.Body, f)
+		}
+	case *Redirection:
+		Walk(n.Left, f)
+		Walk(n.Right, f)
+	case *FunctionDefinition:
+		Walk(n.Name, f)
+		Walk(n.Body, f)
+	case *GroupedExpression:
+		Walk(n.Exp, f)
 	}
 }
+
+func (n *FunctionDefinition) TokenLiteralNode() token.Token      { return n.Token }
+func (n *Redirection) TokenLiteralNode() token.Token             { return n.Token }
+func (n *LetStatement) TokenLiteralNode() token.Token            { return n.Token }
+func (n *ReturnStatement) TokenLiteralNode() token.Token         { return n.Token }
+func (n *ExpressionStatement) TokenLiteralNode() token.Token     { return n.Token }
+func (n *Identifier) TokenLiteralNode() token.Token              { return n.Token }
+func (n *IntegerLiteral) TokenLiteralNode() token.Token          { return n.Token }
+func (n *Boolean) TokenLiteralNode() token.Token                 { return n.Token }
+func (n *PrefixExpression) TokenLiteralNode() token.Token        { return n.Token }
+func (n *PostfixExpression) TokenLiteralNode() token.Token       { return n.Token }
+func (n *InfixExpression) TokenLiteralNode() token.Token         { return n.Token }
+func (n *BlockStatement) TokenLiteralNode() token.Token          { return n.Token }
+func (n *IfStatement) TokenLiteralNode() token.Token             { return n.Token }
+func (n *ForLoopStatement) TokenLiteralNode() token.Token        { return n.Token }
+func (n *WhileLoopStatement) TokenLiteralNode() token.Token      { return n.Token }
+func (n *FunctionLiteral) TokenLiteralNode() token.Token         { return n.Token }
+func (n *CallExpression) TokenLiteralNode() token.Token          { return n.Token }
+func (n *StringLiteral) TokenLiteralNode() token.Token           { return n.Token }
+func (n *BracketExpression) TokenLiteralNode() token.Token       { return n.Token }
+func (n *DoubleBracketExpression) TokenLiteralNode() token.Token { return n.Token }
+func (n *ArrayAccess) TokenLiteralNode() token.Token             { return n.Token }
+func (n *IndexExpression) TokenLiteralNode() token.Token         { return n.Token }
+func (n *InvalidArrayAccess) TokenLiteralNode() token.Token      { return n.Token }
+func (n *CommandSubstitution) TokenLiteralNode() token.Token     { return n.Token }
+func (n *Shebang) TokenLiteralNode() token.Token                 { return n.Token }
+func (n *DollarParenExpression) TokenLiteralNode() token.Token   { return n.Token }
+func (n *SimpleCommand) TokenLiteralNode() token.Token           { return n.Token }
+func (n *ConcatenatedExpression) TokenLiteralNode() token.Token  { return n.Token }
+func (n *CaseStatement) TokenLiteralNode() token.Token           { return n.Token }
+func (n *CaseClause) TokenLiteralNode() token.Token              { return n.Token }
