@@ -1,8 +1,6 @@
 package katas
 
 import (
-	"reflect"
-
 	"github.com/afadesigns/zshellcheck/pkg/ast"
 )
 
@@ -13,11 +11,12 @@ var plumbingCommands = map[string]string{
 }
 
 func init() {
-	RegisterKata(reflect.TypeOf(&ast.SimpleCommand{}), Kata{
-		ID:          "ZC1011",
-		Title:       "Use `git` porcelain commands instead of plumbing commands",
-		Description: "Plumbing commands in `git` are designed for scripting and can be unstable. Porcelain commands are designed for interactive use and are more stable.",
-		Check:       checkZC1011,
+	RegisterKata(ast.SimpleCommandNode, Kata{
+		ID:    "ZC1011",
+		Title: "Use `git` porcelain commands instead of plumbing commands",
+		Description: "Plumbing commands in `git` are designed for scripting and can be unstable. " +
+			"Porcelain commands are designed for interactive use and are more stable.",
+		Check: checkZC1011,
 	})
 }
 
@@ -28,15 +27,14 @@ func checkZC1011(node ast.Node) []Violation {
 		if ident, ok := cmd.Name.(*ast.Identifier); ok {
 			if ident.Value == "git" {
 				for _, arg := range cmd.Arguments {
-					if argIdent, ok := arg.(*ast.Identifier); ok {
-						if _, ok := plumbingCommands[argIdent.Value]; ok {
-							violations = append(violations, Violation{
-								KataID:  "ZC1011",
-								Message: "Avoid using `git` plumbing commands in scripts. They are not guaranteed to be stable.",
-								Line:    ident.Token.Line,
-								Column:  ident.Token.Column,
-							})
-						}
+					val := getArgValue(arg)
+					if _, ok := plumbingCommands[val]; ok {
+						violations = append(violations, Violation{
+							KataID:  "ZC1011",
+							Message: "Avoid using `git` plumbing commands in scripts. They are not guaranteed to be stable.",
+							Line:    ident.Token.Line,
+							Column:  ident.Token.Column,
+						})
 					}
 				}
 			}
@@ -44,4 +42,25 @@ func checkZC1011(node ast.Node) []Violation {
 	}
 
 	return violations
+}
+
+func getArgValue(expr ast.Expression) string {
+	switch e := expr.(type) {
+	case *ast.Identifier:
+		return e.Value
+	case *ast.StringLiteral:
+		return e.Value
+	case *ast.InfixExpression:
+		return getArgValue(e.Left) + e.Operator + getArgValue(e.Right)
+	case *ast.ConcatenatedExpression:
+		var out string
+		for _, p := range e.Parts {
+			out += getArgValue(p)
+		}
+		return out
+	case *ast.PrefixExpression:
+		// Special handling for prefix expressions in command args to avoid parens
+		return e.Operator + getArgValue(e.Right)
+	}
+	return ""
 }
