@@ -1,17 +1,16 @@
 package katas
 
 import (
-	"reflect"
-
 	"github.com/afadesigns/zshellcheck/pkg/ast"
 )
 
 func init() {
-	RegisterKata(reflect.TypeOf(&ast.LetStatement{}), Kata{
-		ID:          "ZC1013",
-		Title:       "Use `((...))` for arithmetic operations instead of `let`",
-		Description: "The `let` command is a shell builtin, but the `((...))` syntax is more portable and generally preferred for arithmetic operations in Zsh.",
-		Check:       checkZC1013,
+	RegisterKata(ast.LetStatementNode, Kata{
+		ID:    "ZC1013",
+		Title: "Use `((...))` for arithmetic operations instead of `let`",
+		Description: "The `let` command is a shell builtin, but the `((...))` syntax is more portable " +
+			"and generally preferred for arithmetic operations in Zsh.",
+		Check: checkZC1013,
 	})
 }
 
