@@ -1,17 +1,17 @@
 package katas
 
 import (
-	"reflect"
-
 	"github.com/afadesigns/zshellcheck/pkg/ast"
 )
 
 func init() {
-	RegisterKata(reflect.TypeOf(&ast.SimpleCommand{}), Kata{
-		ID:          "ZC1020",
-		Title:       "Use `[[ ... ]]` for tests instead of `test`",
-		Description: "The `test` command is an external command and may not be available on all systems. The `[[...]]` construct is a Zsh keyword, offering safer and more powerful conditional expressions than the traditional `test` command.",
-		Check:       checkZC1020,
+	RegisterKata(ast.SimpleCommandNode, Kata{
+		ID:    "ZC1020",
+		Title: "Use `[[ ... ]]` for tests instead of `test`",
+		Description: "The `test` command is an external command and may not be available on all systems. " +
+			"The `[[...]]` construct is a Zsh keyword, offering safer and more powerful conditional " +
+			"expressions than the traditional `test` command.",
+		Check: checkZC1020,
 	})
 }
 
