@@ -1,17 +1,16 @@
 package katas
 
 import (
-	"reflect"
-
 	"github.com/afadesigns/zshellcheck/pkg/ast"
 )
 
 func init() {
-	RegisterKata(reflect.TypeOf(&ast.LetStatement{}), Kata{
-		ID:          "ZC1028",
-		Title:       "Use `$((...))` for arithmetic expansion",
-		Description: "The `$((...))` syntax is the modern, recommended way to perform arithmetic expansion. It is more readable and can be nested easily, unlike `let`.",
-		Check:       checkZC1028,
+	RegisterKata(ast.LetStatementNode, Kata{
+		ID:    "ZC1028",
+		Title: "Use `$((...))` for arithmetic expansion",
+		Description: "The `$((...))` syntax is the modern, recommended way to perform arithmetic expansion. " +
+			"It is more readable and can be nested easily, unlike `let`.",
+		Check: checkZC1028,
 	})
 }
 
