@@ -1,17 +1,16 @@
 package katas
 
 import (
-	"reflect"
-
 	"github.com/afadesigns/zshellcheck/pkg/ast"
 )
 
 func init() {
-	RegisterKata(reflect.TypeOf(&ast.LetStatement{}), Kata{
-		ID:          "ZC1032",
-		Title:       "Use `((...))` for C-style incrementing",
-		Description: "Instead of `let i=i+1`, you can use the more concise and idiomatic C-style increment `(( i++ ))` in Zsh.",
-		Check:       checkZC1032,
+	RegisterKata(ast.LetStatementNode, Kata{
+		ID:    "ZC1032",
+		Title: "Use `((...))` for C-style incrementing",
+		Description: "Instead of `let i=i+1`, you can use the more concise and idiomatic C-style " +
+			"increment `(( i++ ))` in Zsh.",
+		Check: checkZC1032,
 	})
 }
 
@@ -19,9 +18,11 @@ func checkZC1032(node ast.Node) []Violation {
 	violations := []Violation{}
 
 	if letStmt, ok := node.(*ast.LetStatement); ok {
+		// Check for let i = i + 1
 		if infixExpr, ok := letStmt.Value.(*ast.InfixExpression); ok {
 			if leftIdent, ok := infixExpr.Left.(*ast.Identifier); ok {
 				if rightInt, ok := infixExpr.Right.(*ast.IntegerLiteral); ok {
+					// Match name, "+", and 1
 					if letStmt.Name.Value == leftIdent.Value && infixExpr.Operator == "+" && rightInt.Value == 1 {
 						violations = append(violations, Violation{
 							KataID:  "ZC1032",
