@@ -1,17 +1,17 @@
 package katas
 
 import (
-	"reflect"
-
 	"github.com/afadesigns/zshellcheck/pkg/ast"
 )
 
 func init() {
-	RegisterKata(reflect.TypeOf(&ast.SimpleCommand{}), Kata{
-		ID:          "ZC1036",
-		Title:       "Prefer `[[ ... ]]` over `test` command",
-		Description: "The `[[ ... ]]` construct is a more powerful and safer alternative to the `test` command (or `[ ... ]`) for conditional expressions in modern shells. It handles word splitting and globbing more intuitively and supports advanced features like regex matching.",
-		Check:       checkZC1036,
+	RegisterKata(ast.SimpleCommandNode, Kata{
+		ID:    "ZC1036",
+		Title: "Prefer `[[ ... ]]` over `test` command",
+		Description: "The `[[ ... ]]` construct is a more powerful and safer alternative to the `test` " +
+			"command (or `[ ... ]`) for conditional expressions in modern shells. It handles word " +
+			"splitting and globbing more intuitively and supports advanced features like regex matching.",
+		Check: checkZC1036,
 	})
 }
 
