@@ -17,24 +17,28 @@ const (
 	SUM         // +
 	PRODUCT     // *
 	PREFIX      // -X or !X
+	POSTFIX     // X++
 	CALL        // myFunction(X)
-	POSTFIX     // i++
+	INDEX       // array[index]
 )
 
-var precedences = map[token.TokenType]int{
-	token.EQ:       EQUALS,
-	token.NOT_EQ:   EQUALS,
-	token.LT:       LESSGREATER,
-	token.GT:       LESSGREATER,
-	token.PLUS:     SUM,
-	token.MINUS:    SUM,
-	token.SLASH:    PRODUCT,
-	token.ASTERISK: PRODUCT,
-	token.LPAREN:   CALL,
-	token.PIPE:     CALL,
-	token.ASSIGN:   EQUALS,
-	token.INC:      POSTFIX,
-	token.DEC:      POSTFIX,
+var precedences = map[token.Type]int{
+	token.EQ:            EQUALS,
+	token.NotEq:         EQUALS,
+	token.LT:            LESSGREATER,
+	token.GT:            LESSGREATER,
+	token.PLUS:          SUM,
+	token.MINUS:         SUM,
+	token.SLASH:         PRODUCT,
+	token.ASTERISK:      PRODUCT,
+	token.LPAREN:        CALL,
+	token.LBRACKET:      INDEX,
+	token.DollarLbrace:  INDEX,
+	token.DOLLAR_LPAREN: CALL,
+	token.DoubleLparen:  CALL,
+	token.ASSIGN:        EQUALS,
+	token.INC:           POSTFIX,
+	token.DEC:           POSTFIX,
 }
 
 type (
@@ -49,14 +53,16 @@ type Parser struct {
 	curToken  token.Token
 	peekToken token.Token
 
-	prefixParseFns map[token.TokenType]prefixParseFn
-	infixParseFns  map[token.TokenType]infixParseFn
+	prefixParseFns map[token.Type]prefixParseFn
+	infixParseFns  map[token.Type]infixParseFn
+
+	inBackticks int
 }
 
 func New(l *lexer.Lexer) *Parser {
 	p := &Parser{l: l, errors: []string{}}
 
-	p.prefixParseFns = make(map[token.TokenType]prefixParseFn)
+	p.prefixParseFns = make(map[token.Type]prefixParseFn)
 	p.registerPrefix(token.IDENT, p.parseIdentifier)
 	p.registerPrefix(token.INT, p.parseIntegerLiteral)
 	p.registerPrefix(token.BANG, p.parsePrefixExpression)
@@ -66,61 +72,89 @@ func New(l *lexer.Lexer) *Parser {
 	p.registerPrefix(token.LPAREN, p.parseGroupedExpression)
 	p.registerPrefix(token.FUNCTION, p.parseFunctionLiteral)
 	p.registerPrefix(token.STRING, p.parseStringLiteral)
-	p.registerPrefix(token.LBRACKET, p.parseBracketExpression)
+	p.registerPrefix(token.LBRACKET, p.parseSingleCommand)
 	p.registerPrefix(token.LDBRACKET, p.parseDoubleBracketExpression)
-	p.registerPrefix(token.DOLLAR_LBRACE, p.parseArrayAccess)
+	p.registerPrefix(token.DollarLbrace, p.parseArrayAccess)
 	p.registerPrefix(token.DOLLAR, p.parseInvalidArrayAccessPrefix)
+	p.registerPrefix(token.VARIABLE, p.parseIdentifier)
 	p.registerPrefix(token.DOLLAR_LPAREN, p.parseDollarParenExpression)
-	p.registerPrefix(token.DOUBLE_LPAREN, p.parseDoubleParenExpression)
+	p.registerPrefix(token.DoubleLparen, p.parseDoubleParenExpression)
 	p.registerPrefix(token.BACKTICK, p.parseCommandSubstitution)
 
-	p.infixParseFns = make(map[token.TokenType]infixParseFn)
+	p.infixParseFns = make(map[token.Type]infixParseFn)
 	p.registerInfix(token.PLUS, p.parseInfixExpression)
 	p.registerInfix(token.MINUS, p.parseInfixExpression)
 	p.registerInfix(token.SLASH, p.parseInfixExpression)
 	p.registerInfix(token.ASTERISK, p.parseInfixExpression)
+	p.registerInfix(token.AND, p.parseInfixExpression)
+	p.registerInfix(token.OR, p.parseInfixExpression)
 	p.registerInfix(token.EQ, p.parseInfixExpression)
-	p.registerInfix(token.NOT_EQ, p.parseInfixExpression)
+	p.registerInfix(token.NotEq, p.parseInfixExpression)
 	p.registerInfix(token.LT, p.parseInfixExpression)
 	p.registerInfix(token.GT, p.parseInfixExpression)
 	p.registerInfix(token.LPAREN, p.parseCallExpression)
-	p.registerInfix(token.PIPE, p.parseInfixExpression)
+	p.registerInfix(token.LBRACKET, p.parseIndexExpression)
 	p.registerInfix(token.ASSIGN, p.parseInfixExpression)
 	p.registerInfix(token.INC, p.parsePostfixExpression)
 	p.registerInfix(token.DEC, p.parsePostfixExpression)
+	p.registerInfix(token.GTGT, p.parseRedirection)
+	p.registerInfix(token.LTLT, p.parseRedirection)
+	p.registerInfix(token.GTAMP, p.parseRedirection)
+	p.registerInfix(token.LTAMP, p.parseRedirection)
 
-	p.nextToken()
-	p.nextToken()
+	p.nextToken() // Initialize curToken
+	p.nextToken() // Initialize peekToken
 
 	return p
+}
+
+func (p *Parser) parseRedirection(left ast.Expression) ast.Expression {
+	expr := &ast.Redirection{
+		Token:    p.curToken,
+		Operator: p.curToken.Literal,
+		Left:     left,
+	}
+
+	p.nextToken()
+	expr.Right = p.parseExpression(LOWEST)
+
+	return expr
+}
+
+func (p *Parser) parseIndexExpression(left ast.Expression) ast.Expression {
+	exp := &ast.IndexExpression{Token: p.curToken, Left: left}
+
+	p.nextToken()
+	exp.Index = p.parseExpression(LOWEST)
+
+	if !p.expectPeek(token.RBRACKET) {
+		return nil
+	}
+
+	return exp
 }
 
 func (p *Parser) Errors() []string {
 	return p.errors
 }
 
-func (p *Parser) nextToken() {
-	p.curToken = p.peekToken
-	p.peekToken = p.l.NextToken()
-}
-
 func (p *Parser) ParseProgram() *ast.Program {
 	program := &ast.Program{}
 	program.Statements = []ast.Statement{}
 
-	for p.curToken.Type != token.EOF {
-		if p.curTokenIs(token.SEMICOLON) {
-			p.nextToken()
-			continue
-		}
+	for !p.curTokenIs(token.EOF) {
 		stmt := p.parseStatement()
 		if stmt != nil {
 			program.Statements = append(program.Statements, stmt)
 		}
 		p.nextToken()
 	}
-
 	return program
+}
+
+func (p *Parser) nextToken() {
+	p.curToken = p.peekToken
+	p.peekToken = p.l.NextToken()
 }
 
 func (p *Parser) parseStatement() ast.Statement {
@@ -133,16 +167,250 @@ func (p *Parser) parseStatement() ast.Statement {
 		return p.parseLetStatement()
 	case token.RETURN:
 		return p.parseReturnStatement()
-	case token.IF:
+	case token.If:
 		return p.parseIfStatement()
 	case token.SHEBANG:
 		return p.parseShebangStatement()
+	case token.HASH:
+		// Skip comments for now
+		return nil
 	case token.FOR:
 		return p.parseForLoopStatement()
+	case token.WHILE:
+		return p.parseWhileLoopStatement()
+	case token.LBRACE:
+		tok := p.curToken
+		p.nextToken()
+		block := p.parseBlockStatement(token.RBRACE)
+		block.Token = tok
+		return block
+	case token.LPAREN:
+		return p.parseSubshellStatement()
+	case token.COLON, token.DOT, token.LBRACKET:
+		return p.parseSimpleCommandStatement()
+	case token.CASE:
+		return p.parseCaseStatement()
+	case token.IDENT:
+		if p.curToken.Literal == "test" {
+			return p.parseSimpleCommandStatement()
+		}
+		if p.peekTokenIs(token.IDENT) || p.peekTokenIs(token.STRING) || p.peekTokenIs(token.INT) ||
+			p.peekTokenIs(token.MINUS) || p.peekTokenIs(token.DOT) || p.peekTokenIs(token.VARIABLE) ||
+			p.peekTokenIs(token.DOLLAR) || p.peekTokenIs(token.DollarLbrace) ||
+			p.peekTokenIs(token.DOLLAR_LPAREN) || p.peekTokenIs(token.SLASH) ||
+			p.peekTokenIs(token.TILDE) || p.peekTokenIs(token.ASTERISK) ||
+			p.peekTokenIs(token.BANG) {
+			return p.parseSimpleCommandStatement()
+		}
+		return p.parseExpressionOrFunctionDefinition()
 	default:
-		return p.parseExpressionStatement()
+		return p.parseExpressionOrFunctionDefinition()
 	}
-	return nil
+}
+
+func (p *Parser) parseExpressionOrFunctionDefinition() ast.Statement {
+	stmt := p.parseExpressionStatement()
+
+	// Check if it matches function definition pattern: name()
+	if call, ok := stmt.Expression.(*ast.CallExpression); ok {
+		if len(call.Arguments) == 0 {
+			if ident, ok := call.Function.(*ast.Identifier); ok {
+				// It is `name()`. In Zsh this must be followed by a body to be a valid definition.
+				// If we are here, we consumed `name` and `()`.
+				// Parse the next statement as body.
+				funcDef := &ast.FunctionDefinition{
+					Token: ident.Token,
+					Name:  ident,
+				}
+				// We expect a statement now.
+				// If we are at semicolon, skip it? `func(); body` is valid? No. `func() body`.
+				// But lexer might have produced semicolon if newline?
+				// If next is `{`, `(` (subshell body), or command.
+
+				// If we are at EOF or semicolon without body, it's just a CallExpression (incomplete func def).
+				if p.curTokenIs(token.SEMICOLON) || p.curTokenIs(token.EOF) {
+					return stmt
+				}
+
+				// Parse body
+				p.nextToken() // Move to start of body statement
+				funcDef.Body = p.parseStatement()
+				return funcDef
+			}
+		}
+	}
+	return stmt
+}
+
+func (p *Parser) parseSimpleCommandStatement() ast.Statement {
+	stmt := &ast.ExpressionStatement{Token: p.curToken}
+
+	stmt.Expression = p.parseCommandList()
+
+	if p.peekTokenIs(token.SEMICOLON) {
+		p.nextToken()
+	}
+
+	return stmt
+}
+
+func (p *Parser) parseCommandList() ast.Expression {
+	left := p.parseCommandPipeline()
+
+	for p.peekTokenIs(token.AND) || p.peekTokenIs(token.OR) {
+		p.nextToken()
+		op := p.curToken
+		p.nextToken() // move to start of right command
+		right := p.parseCommandPipeline()
+		left = &ast.InfixExpression{
+			Token:    op,
+			Operator: op.Literal,
+			Left:     left,
+			Right:    right,
+		}
+	}
+
+	return left
+}
+
+func (p *Parser) parseCommandPipeline() ast.Expression {
+	if p.curTokenIs(token.BANG) {
+		tok := p.curToken
+		p.nextToken()
+		right := p.parseCommandPipeline()
+		return &ast.PrefixExpression{Token: tok, Operator: "!", Right: right}
+	}
+
+	left := p.parseSingleCommand()
+
+	// Parse redirections
+	for p.peekTokenIs(token.GT) || p.peekTokenIs(token.GTGT) ||
+		p.peekTokenIs(token.LT) || p.peekTokenIs(token.LTLT) ||
+		p.peekTokenIs(token.GTAMP) || p.peekTokenIs(token.LTAMP) {
+
+		p.nextToken()
+		op := p.curToken
+		p.nextToken() // consume op
+		// Redirection target is file/expression. Use parseCommandWord to handle paths/strings correctly.
+		right := p.parseCommandWord()
+
+		left = &ast.Redirection{
+			Token:    op,
+			Left:     left,
+			Operator: op.Literal,
+			Right:    right,
+		}
+	}
+
+	if p.peekTokenIs(token.PIPE) {
+		p.nextToken() // consume '|'
+		op := p.curToken
+		p.nextToken() // move to the start of the next command
+		right := p.parseCommandPipeline()
+		left = &ast.InfixExpression{
+			Token:    op,
+			Operator: op.Literal,
+			Left:     left,
+			Right:    right,
+		}
+	}
+
+	return left
+}
+
+func (p *Parser) isCommandDelimiter(t token.Token) bool {
+	if t.Type == token.BACKTICK && p.inBackticks > 0 {
+		return true // Terminating backtick is a delimiter
+	}
+	// If NOT in backticks, start backtick is NOT a delimiter (it starts a substitution)
+	if t.Type == token.BACKTICK && p.inBackticks == 0 {
+		return false
+	}
+
+	return t.Type == token.EOF || t.Type == token.SEMICOLON || t.Type == token.PIPE ||
+		t.Type == token.AND || t.Type == token.OR ||
+		t.Type == token.RPAREN || t.Type == token.RBRACE ||
+		t.Type == token.HASH ||
+		t.Type == token.THEN || t.Type == token.ELSE || t.Type == token.ELIF || t.Type == token.Fi ||
+		t.Type == token.DO || t.Type == token.DONE ||
+		t.Type == token.ESAC || t.Type == token.DSEMI ||
+		t.Type == token.GT || t.Type == token.LT ||
+		t.Type == token.GTGT || t.Type == token.LTLT ||
+		t.Type == token.GTAMP || t.Type == token.LTAMP
+}
+
+func (p *Parser) parseSingleCommand() ast.Expression {
+	cmd := &ast.SimpleCommand{
+		Token: p.curToken,
+		Name:  &ast.Identifier{Token: p.curToken, Value: p.curToken.Literal},
+	}
+
+	// Check for function definition syntax: name()
+	if p.peekTokenIs(token.LPAREN) {
+		p.nextToken() // curToken is (
+		if p.peekTokenIs(token.RPAREN) {
+			p.nextToken() // curToken is )
+			// Function Definition!
+			funcDef := &ast.FunctionDefinition{
+				Token: cmd.Token,
+				Name:  cmd.Name.(*ast.Identifier),
+			}
+
+			p.nextToken() // Move to start of body
+			funcDef.Body = p.parseStatement()
+			return funcDef
+		} else {
+			// It was not (), it was `name ( ...`
+			arg := p.parseCommandWord()
+			cmd.Arguments = append(cmd.Arguments, arg)
+		}
+	} else {
+		cmd.Arguments = []ast.Expression{}
+	}
+
+	// Continue parsing arguments
+	for !p.isCommandDelimiter(p.peekToken) && p.peekToken.Line == p.curToken.Line {
+		p.nextToken()
+		arg := p.parseCommandWord()
+		cmd.Arguments = append(cmd.Arguments, arg)
+	}
+
+	return cmd
+}
+
+func (p *Parser) parseCommandWord() ast.Expression {
+	firstToken := p.curToken
+	parts := []ast.Expression{}
+
+	// Parse the first part
+	if p.prefixParseFns[p.curToken.Type] == nil {
+		parts = append(parts, &ast.StringLiteral{Token: p.curToken, Value: p.curToken.Literal})
+	} else {
+		parts = append(parts, p.parseExpression(CALL))
+	}
+
+	// Continue parsing while the next token is adjacent (no preceding space)
+	for !p.peekToken.HasPrecedingSpace && !p.isCommandDelimiter(p.peekToken) &&
+		p.peekToken.Line == p.curToken.Line {
+
+		p.nextToken()
+
+		if p.prefixParseFns[p.curToken.Type] == nil {
+			// Treat as literal string part
+			parts = append(parts, &ast.StringLiteral{Token: p.curToken, Value: p.curToken.Literal})
+		} else {
+			parts = append(parts, p.parseExpression(CALL))
+		}
+	}
+
+	if len(parts) == 1 {
+		return parts[0]
+	}
+
+	return &ast.ConcatenatedExpression{
+		Token: firstToken,
+		Parts: parts,
+	}
 }
 
 func (p *Parser) parseLetStatement() *ast.LetStatement {
@@ -186,45 +454,52 @@ func (p *Parser) parseExpressionStatement() *ast.ExpressionStatement {
 func (p *Parser) parseIfStatement() *ast.IfStatement {
 	stmt := &ast.IfStatement{Token: p.curToken}
 	p.nextToken()
-	stmt.Condition = p.parseExpression(LOWEST)
-	if p.peekTokenIs(token.SEMICOLON) {
-		p.nextToken()
-	}
-	if !p.expectPeek(token.THEN) {
+	stmt.Condition = p.parseBlockStatement(token.THEN)
+
+	if !p.curTokenIs(token.THEN) {
 		return nil
 	}
+
 	p.nextToken() // consume "then"
-	stmt.Consequence = p.parseBlockStatement(token.ELSE, token.FI)
+	stmt.Consequence = p.parseBlockStatement(token.ELSE, token.Fi)
 
 	if p.curTokenIs(token.ELSE) {
 		p.nextToken() // consume "else"
-		stmt.Alternative = p.parseBlockStatement(token.FI)
+		stmt.Alternative = p.parseBlockStatement(token.Fi)
 	}
-	if !p.curTokenIs(token.FI) {
-		p.peekError(token.FI)
+	if !p.curTokenIs(token.Fi) {
+		p.peekError(token.Fi)
 		return nil
 	}
 	return stmt
 }
 
-func (p *Parser) parseBlockStatement(terminators ...token.TokenType) *ast.BlockStatement {
+func (p *Parser) parseBlockStatement(terminators ...token.Type) *ast.BlockStatement {
 	block := &ast.BlockStatement{Token: p.curToken}
 	block.Statements = []ast.Statement{}
 
-	isTerminator := func(tok token.Token) bool {
+	for !p.curTokenIs(token.EOF) {
+		isTerm := false
 		for _, t := range terminators {
-			if tok.Type == t {
-				return true
+			if p.curTokenIs(t) {
+				isTerm = true
+				break
 			}
 		}
-		return false
-	}
-
-	for !isTerminator(p.curToken) && !p.curTokenIs(token.EOF) {
-		s := p.parseStatement()
-		if s != nil {
-			block.Statements = append(block.Statements, s)
+		if isTerm {
+			break
 		}
+
+		if p.curTokenIs(token.SEMICOLON) {
+			p.nextToken()
+			continue
+		}
+
+		stmt := p.parseStatement()
+		if stmt != nil {
+			block.Statements = append(block.Statements, stmt)
+		}
+
 		p.nextToken()
 	}
 	return block
@@ -237,6 +512,7 @@ func (p *Parser) parseExpression(precedence int) ast.Expression {
 		return nil
 	}
 	leftExp := prefix()
+
 	for !p.peekTokenIs(token.SEMICOLON) && precedence < p.peekPrecedence() {
 		infix := p.infixParseFns[p.peekToken.Type]
 		if infix == nil {
@@ -286,26 +562,6 @@ func (p *Parser) parsePostfixExpression(left ast.Expression) ast.Expression {
 	return &ast.PostfixExpression{Token: p.curToken, Left: left, Operator: p.curToken.Literal}
 }
 
-func (p *Parser) parseBracketExpression() ast.Expression {
-	bracketToken := p.curToken
-	p.nextToken()
-	expressions := []ast.Expression{}
-	for !p.curTokenIs(token.RBRACKET) && !p.curTokenIs(token.EOF) {
-		exp := p.parseExpression(LOWEST)
-		if exp != nil {
-			expressions = append(expressions, exp)
-		}
-		if !p.curTokenIs(token.RBRACKET) && !p.curTokenIs(token.EOF) {
-			p.nextToken()
-		}
-	}
-	if !p.curTokenIs(token.RBRACKET) {
-		p.peekError(token.RBRACKET)
-		return nil
-	}
-	return &ast.BracketExpression{Token: bracketToken, Expressions: expressions}
-}
-
 func (p *Parser) parseDoubleBracketExpression() ast.Expression {
 	bracketToken := p.curToken
 	p.nextToken()
@@ -339,12 +595,13 @@ func (p *Parser) parseInfixExpression(left ast.Expression) ast.Expression {
 }
 
 func (p *Parser) parseGroupedExpression() ast.Expression {
+	tok := p.curToken
 	p.nextToken()
 	exp := p.parseExpression(LOWEST)
 	if !p.expectPeek(token.RPAREN) {
 		return nil
 	}
-	return exp
+	return &ast.GroupedExpression{Token: tok, Exp: exp}
 }
 
 func (p *Parser) parseArrayAccess() ast.Expression {
@@ -353,73 +610,133 @@ func (p *Parser) parseArrayAccess() ast.Expression {
 		return nil
 	}
 	exp.Left = p.parseIdentifier()
-	if !p.expectPeek(token.LBRACKET) {
-		return nil
+
+	// check for optional index
+	if p.peekTokenIs(token.LBRACKET) {
+		p.nextToken() // consume [
+		p.nextToken() // move to start of index expression
+		exp.Index = p.parseExpression(LOWEST)
+		if !p.expectPeek(token.RBRACKET) {
+			return nil
+		}
 	}
-	p.nextToken()
-	exp.Index = p.parseExpression(LOWEST)
-	if !p.expectPeek(token.RBRACKET) {
-		return nil
-	}
+
 	if !p.expectPeek(token.RBRACE) {
 		return nil
 	}
+
 	return exp
 }
 
 func (p *Parser) parseInvalidArrayAccessPrefix() ast.Expression {
-    dollarToken := p.curToken
-    if !p.expectPeek(token.IDENT) {
-        return nil
-    }
-    
-    ident := &ast.Identifier{Token: p.curToken, Value: p.curToken.Literal}
+	dollarToken := p.curToken
+	if !p.expectPeek(token.IDENT) {
+		return nil
+	}
 
-    if !p.peekTokenIs(token.LBRACKET) {
-        return &ast.PrefixExpression{Token: dollarToken, Operator: "$", Right: ident}
-    }
+	ident := &ast.Identifier{Token: p.curToken, Value: p.curToken.Literal}
 
-    p.nextToken()
+	if !p.peekTokenIs(token.LBRACKET) {
+		return &ast.PrefixExpression{Token: dollarToken, Operator: "$", Right: ident}
+	}
 
-    exp := &ast.InvalidArrayAccess{Token: dollarToken, Left: ident}
+	p.nextToken()
 
-    p.nextToken()
-    exp.Index = p.parseExpression(LOWEST)
+	exp := &ast.InvalidArrayAccess{Token: dollarToken, Left: ident}
 
-    if !p.expectPeek(token.RBRACKET) {
-        return nil
-    }
+	p.nextToken()
+	exp.Index = p.parseExpression(LOWEST)
 
-    return exp
+	if !p.expectPeek(token.RBRACKET) {
+		return nil
+	}
+
+	return exp
+}
+
+func (p *Parser) parseSubshellStatement() ast.Statement {
+	subshellToken := p.curToken
+	p.nextToken()
+	block := p.parseBlockStatement(token.RPAREN)
+	if !p.curTokenIs(token.RPAREN) {
+		p.peekError(token.RPAREN)
+		return nil
+	}
+	p.nextToken()
+	block.Token = subshellToken
+	return block
+}
+
+func (p *Parser) parseCaseStatement() *ast.CaseStatement {
+	stmt := &ast.CaseStatement{Token: p.curToken}
+	p.nextToken()
+	stmt.Value = p.parseExpression(LOWEST)
+
+	if !p.expectPeek(token.IN) {
+		return nil
+	}
+	p.nextToken()
+
+	for !p.curTokenIs(token.ESAC) && !p.curTokenIs(token.EOF) {
+		for p.curTokenIs(token.SEMICOLON) {
+			p.nextToken()
+		}
+		if p.curTokenIs(token.ESAC) {
+			break
+		}
+		clause := &ast.CaseClause{Token: p.curToken}
+		if p.curTokenIs(token.LPAREN) {
+			p.nextToken()
+		}
+		for {
+			pat := p.parseCommandWord()
+			clause.Patterns = append(clause.Patterns, pat)
+			if p.peekTokenIs(token.PIPE) {
+				p.nextToken()
+				p.nextToken()
+			} else {
+				break
+			}
+		}
+		if !p.expectPeek(token.RPAREN) {
+			return nil
+		}
+		p.nextToken()
+		clause.Body = p.parseBlockStatement(token.DSEMI, token.ESAC)
+		stmt.Clauses = append(stmt.Clauses, clause)
+		if p.curTokenIs(token.DSEMI) {
+			p.nextToken()
+		}
+	}
+	return stmt
 }
 
 func (p *Parser) parseFunctionLiteral() ast.Expression {
 	lit := &ast.FunctionLiteral{Token: p.curToken}
-
-	// Zsh functions can be anonymous `function() {}` or named `function my_func() {}`
 	if p.peekTokenIs(token.IDENT) {
-		p.nextToken() // consume function name
+		p.nextToken()
+		lit.Name = p.curToken.Literal
 	}
-
 	if !p.expectPeek(token.LPAREN) {
 		return nil
 	}
-
 	lit.Parameters = p.parseFunctionParameters()
-
 	if !p.expectPeek(token.LBRACE) {
 		return nil
 	}
-	p.nextToken() // consume "{"
+	p.nextToken()
 	lit.Body = p.parseBlockStatement(token.RBRACE)
-
 	return lit
 }
 
 func (p *Parser) parseCommandSubstitution() ast.Expression {
 	exp := &ast.CommandSubstitution{Token: p.curToken}
 	p.nextToken()
-	exp.Command = p.parseExpression(LOWEST)
+
+	p.inBackticks++
+	exp.Command = p.parseCommandList()
+	p.inBackticks--
+
 	if !p.expectPeek(token.BACKTICK) {
 		return nil
 	}
@@ -428,8 +745,30 @@ func (p *Parser) parseCommandSubstitution() ast.Expression {
 
 func (p *Parser) parseDollarParenExpression() ast.Expression {
 	exp := &ast.DollarParenExpression{Token: p.curToken}
+
+	if p.peekTokenIs(token.LPAREN) {
+		p.nextToken()
+		p.nextToken() // consume `(`
+		cmd := p.parseExpression(LOWEST)
+
+		if p.peekTokenIs(token.DoubleRparen) {
+			p.nextToken() // consume ))
+			exp.Command = cmd
+			return exp
+		}
+
+		if !p.expectPeek(token.RPAREN) {
+			return nil
+		}
+		if !p.expectPeek(token.RPAREN) {
+			return nil
+		}
+		exp.Command = cmd
+		return exp
+	}
+
 	p.nextToken()
-	exp.Command = p.parseExpression(LOWEST)
+	exp.Command = p.parseCommandList()
 	if !p.expectPeek(token.RPAREN) {
 		return nil
 	}
@@ -482,15 +821,6 @@ func (p *Parser) parseCallArguments() []ast.Expression {
 	return args
 }
 
-func (p *Parser) parseCommandArguments() []ast.Expression {
-	args := []ast.Expression{}
-	for !p.curTokenIs(token.SEMICOLON) && !p.curTokenIs(token.EOF) {
-		args = append(args, p.parseExpression(LOWEST))
-		p.nextToken()
-	}
-	return args
-}
-
 func (p *Parser) parseShebangStatement() *ast.Shebang {
 	return &ast.Shebang{Token: p.curToken, Path: p.curToken.Literal}
 }
@@ -498,70 +828,126 @@ func (p *Parser) parseShebangStatement() *ast.Shebang {
 func (p *Parser) parseForLoopStatement() *ast.ForLoopStatement {
 	stmt := &ast.ForLoopStatement{Token: p.curToken}
 
-	if !p.expectPeek(token.DOUBLE_LPAREN) {
-		return nil
+	if p.peekTokenIs(token.DoubleLparen) {
+		// Arithmetic for loop: for (( init; cond; post ))
+		p.nextToken() // consume ((
+
+		// Init (optional)
+		if !p.peekTokenIs(token.SEMICOLON) {
+			p.nextToken()
+			if p.prefixParseFns[p.curToken.Type] != nil {
+				stmt.Init = p.parseExpression(LOWEST)
+			} else {
+				p.noPrefixParseFnError(p.curToken.Type)
+				return nil
+			}
+		}
+		if !p.expectPeek(token.SEMICOLON) {
+			return nil
+		}
+
+		// Condition (optional)
+		if !p.peekTokenIs(token.SEMICOLON) {
+			p.nextToken()
+			if p.prefixParseFns[p.curToken.Type] != nil {
+				stmt.Condition = p.parseExpression(LOWEST)
+			} else {
+				p.noPrefixParseFnError(p.curToken.Type)
+				return nil
+			}
+		}
+		if !p.expectPeek(token.SEMICOLON) {
+			return nil
+		}
+
+		// Post (optional)
+		if !p.peekTokenIs(token.DoubleRparen) {
+			p.nextToken()
+			if p.prefixParseFns[p.curToken.Type] != nil {
+				stmt.Post = p.parseExpression(LOWEST)
+			} else {
+				p.noPrefixParseFnError(p.curToken.Type)
+				return nil
+			}
+		}
+
+		if !p.expectPeek(token.DoubleRparen) {
+			return nil
+		}
+
+		// Optional semicolon before DO
+		if p.peekTokenIs(token.SEMICOLON) {
+			p.nextToken()
+		}
+
+		if !p.expectPeek(token.DO) {
+			return nil
+		}
+		p.nextToken() // consume DO
+		stmt.Body = p.parseBlockStatement(token.DONE)
+		return stmt
 	}
 
-	p.nextToken()
-	stmt.Init = p.parseExpression(LOWEST)
-	p.nextToken()
-
-	if !p.curTokenIs(token.SEMICOLON) {
+	// For-each loop: for name [in words]; do
+	if !p.expectPeek(token.IDENT) {
 		return nil
 	}
-	p.nextToken()
+	stmt.Init = &ast.Identifier{Token: p.curToken, Value: p.curToken.Literal}
 
-	stmt.Condition = p.parseExpression(LOWEST)
-	p.nextToken()
-
-	if !p.curTokenIs(token.SEMICOLON) {
-		return nil
-	}
-	p.nextToken()
-
-	stmt.Post = p.parseExpression(LOWEST)
-
-	if !p.expectPeek(token.RPAREN) {
-		return nil
-	}
-	if !p.expectPeek(token.RPAREN) {
-		return nil
-	}
-	if !p.expectPeek(token.SEMICOLON) {
-		return nil
+	if p.peekTokenIs(token.IN) {
+		p.nextToken()
+		stmt.Items = []ast.Expression{}
+		for !p.peekTokenIs(token.SEMICOLON) && !p.peekTokenIs(token.DO) && !p.peekTokenIs(token.EOF) &&
+			p.peekToken.Line == p.curToken.Line {
+			p.nextToken()
+			arg := p.parseCommandWord()
+			stmt.Items = append(stmt.Items, arg)
+		}
 	}
 
-	if !p.expectPeek(token.IDENT) || p.curToken.Literal != "do" {
+	if p.peekTokenIs(token.SEMICOLON) {
+		p.nextToken()
+	}
+
+	if !p.expectPeek(token.DO) {
 		return nil
 	}
 
 	p.nextToken()
 	stmt.Body = p.parseBlockStatement(token.DONE)
+	return stmt
+}
 
+func (p *Parser) parseWhileLoopStatement() *ast.WhileLoopStatement {
+	stmt := &ast.WhileLoopStatement{Token: p.curToken}
+	p.nextToken()
+	stmt.Condition = p.parseBlockStatement(token.DO)
+	if !p.curTokenIs(token.DO) {
+		return nil
+	}
+	p.nextToken()
+	stmt.Body = p.parseBlockStatement(token.DONE)
 	return stmt
 }
 
 func (p *Parser) parseDoubleParenExpression() ast.Expression {
 	p.nextToken()
 	exp := p.parseExpression(LOWEST)
-	if !p.expectPeek(token.RPAREN) {
-		return nil
-	}
-	if !p.expectPeek(token.RPAREN) {
+	if !p.expectPeek(token.DoubleRparen) {
 		return nil
 	}
 	return exp
 }
 
-func (p *Parser) curTokenIs(t token.TokenType) bool {
+func (p *Parser) curTokenIs(t token.Type) bool {
 	return p.curToken.Type == t
 }
 
-func (p *Parser) peekTokenIs(t token.TokenType) bool {
+func (p *Parser) peekTokenIs(t token.Type) bool {
 	return p.peekToken.Type == t
 }
 
-func (p *Parser) expectPeek(t token.TokenType) bool {
+func (p *Parser) expectPeek(t token.Type) bool {
 	if p.peekTokenIs(t) {
 		p.nextToken()
 		return true
@@ -570,21 +956,21 @@ func (p *Parser) expectPeek(t token.TokenType) bool {
 	return false
 }
 
-func (p *Parser) peekError(t token.TokenType) {
+func (p *Parser) peekError(t token.Type) {
 	msg := fmt.Sprintf("expected next token to be %s, got %s instead", t, p.peekToken.Type)
 	p.errors = append(p.errors, msg)
 }
 
-func (p *Parser) noPrefixParseFnError(t token.TokenType) {
+func (p *Parser) noPrefixParseFnError(t token.Type) {
 	msg := fmt.Sprintf("no prefix parse function for %s found", t)
 	p.errors = append(p.errors, msg)
 }
 
-func (p *Parser) registerPrefix(tokenType token.TokenType, fn prefixParseFn) {
+func (p *Parser) registerPrefix(tokenType token.Type, fn prefixParseFn) {
 	p.prefixParseFns[tokenType] = fn
 }
 
-func (p *Parser) registerInfix(tokenType token.TokenType, fn infixParseFn) {
+func (p *Parser) registerInfix(tokenType token.Type, fn infixParseFn) {
 	p.infixParseFns[tokenType] = fn
 }
 
